Move all the test methods to corresponding modules
