+++ moved/src/move_execution/tests/mod.rs
@@ -1,0 +1,61 @@
+pub use utils::*;
+
+use {
+    super::*,
+    crate::{
+        block::HeaderForExecution,
+        genesis::{config::CHAIN_ID, init_state, L2_CROSS_DOMAIN_MESSENGER_ADDRESS},
+        move_execution::eth_token::quick_get_eth_balance,
+        primitives::{ToMoveAddress, ToMoveU256, B256, U256, U64},
+        storage::{InMemoryState, State},
+        tests::{signer::Signer, ALT_EVM_ADDRESS, ALT_PRIVATE_KEY, EVM_ADDRESS, PRIVATE_KEY},
+        types::transactions::{DepositedTx, ExtendedTxEnvelope, ScriptOrModule},
+    },
+    alloy::{
+        consensus::{transaction::TxEip1559, SignableTransaction, TxEnvelope},
+        network::TxSignerSync,
+        primitives::{address, hex, keccak256, Address, Bytes, FixedBytes, TxKind},
+        rlp::Encodable,
+    },
+    anyhow::Context,
+    aptos_types::{
+        contract_event::ContractEventV2,
+        transaction::{EntryFunction, Module, Script, TransactionArgument},
+    },
+    move_binary_format::{
+        file_format::{
+            AbilitySet, FieldDefinition, IdentifierIndex, ModuleHandleIndex, SignatureToken,
+            StructDefinition, StructFieldInformation, StructHandle, StructHandleIndex,
+            TypeSignature,
+        },
+        CompiledModule,
+    },
+    move_compiler::{
+        shared::{NumberFormat, NumericalAddress},
+        Compiler, Flags,
+    },
+    move_core_types::{
+        account_address::AccountAddress,
+        identifier::Identifier,
+        language_storage::{ModuleId, StructTag},
+        resolver::ModuleResolver,
+        value::{MoveStruct, MoveValue},
+    },
+    move_vm_runtime::module_traversal::{TraversalContext, TraversalStorage},
+    move_vm_types::gas::UnmeteredGasMeter,
+    serde::de::DeserializeOwned,
+    std::{
+        collections::{BTreeMap, BTreeSet},
+        path::Path,
+    },
+};
+
+mod counter;
+mod data_type;
+mod framework;
+mod l1_cost;
+mod marketplace;
+mod natives;
+mod transaction;
+mod transfer;
+mod utils;
