--- moved/src/move_execution/tests2/mod.rs
@@ -1,41 +1,0 @@
-pub use utils::*;
-
-use {
-    super::*,
-    crate::{
-        block::HeaderForExecution,
-        genesis::{config::CHAIN_ID, init_state},
-        move_execution::eth_token::quick_get_eth_balance,
-        primitives::{ToMoveAddress, ToMoveU256, B256, U256, U64},
-        storage::{InMemoryState, State},
-        tests::{signer::Signer, EVM_ADDRESS, PRIVATE_KEY},
-        types::transactions::{DepositedTx, ExtendedTxEnvelope, ScriptOrModule},
-    },
-    alloy::{
-        consensus::{transaction::TxEip1559, SignableTransaction, TxEnvelope},
-        network::TxSignerSync,
-        primitives::{keccak256, Address, FixedBytes, TxKind},
-        rlp::Encodable,
-    },
-    anyhow::Context,
-    aptos_types::transaction::{EntryFunction, Module},
-    move_compiler::{
-        shared::{NumberFormat, NumericalAddress},
-        Compiler, Flags,
-    },
-    move_core_types::{
-        account_address::AccountAddress,
-        identifier::Identifier,
-        language_storage::{ModuleId, StructTag},
-        resolver::ModuleResolver,
-        value::MoveValue,
-    },
-    serde::de::DeserializeOwned,
-    std::{
-        collections::{BTreeMap, BTreeSet},
-        path::Path,
-    },
-};
-
-mod counter_tests;
-mod utils;
